+++ docs/Changelog.md
@@ -4,6 +4,7 @@ title: Changelog
 
 ## Not released yet
 
+* 2016-12-26 Most basic new command ever
 * 2016-12-26 Remove hard dependency on git, login is stored in db now
 * 2016-12-17 `journal` field for posts to publish to specific communities